+++ UnlimitedTowerContract/Contract_eosiocpp/Rule/rule_item_tier_table.hpp
@@ -3,16 +3,16 @@
 
 struct item_stat
 {
-    uint8_t i_str;
-    uint8_t i_dex;
-    uint8_t i_int;
+    uint32_t i_str;
+    uint32_t i_dex;
+    uint32_t i_int;
 };
 //@abi table critemtier i64
 class critemtier
 {
   public:
     uint64_t i_tier;
-    uint8_t i_level;
+    uint32_t i_level;
     item_stat i_min_range;
     item_stat i_max_range;
   public:
