+++ UnlimitedTowerContract/Contract_eosiocpp/Rule/rule_monster_grade_table.hpp
@@ -4,20 +4,20 @@
 
 struct monster_stat
 {
-    uint8_t m_str;
-    uint8_t m_dex;
-    uint8_t m_int;
+    uint32_t m_str;
+    uint32_t m_dex;
+    uint32_t m_int;
 };
 
 //@abi table crmgrade i64
 class crmgrade
 {
   public:
-    uint16_t m_grade;
+    uint64_t m_grade;
     monster_stat m_min_range;
     monster_stat m_max_range;
 public:
-    uint16_t primary_key() const {return m_grade;}
+    uint64_t primary_key() const {return m_grade;}
 };
 
 typedef multi_index<N(crmgrade),crmgrade> monster_grade_data_table;
